refactor(web): single-source the role vocabulary in util/roles (P5)
Adding a role was a ~8-file change scattered across resolveRole, teamRoster,
and their consumers. Introduce util/roles as the one home for the derived
product role vocabulary — RosterRole/EffectiveRole/OrgRole/ViewAsRole/
ClassroomTeamRole, the (previously duplicated) ROLE_RANK + sortRolesByRank,
and the admin<->owner mapping (orgRoleForRole/roleForOrgRole) whose two
directions now sit together. The one contract-frozen literal, StaffRole,
stays in types/classroom.ts (it mirrors the persisted teams schema); every
union derives from it, so a new role flows in from ~1-2 files.

resolveRole and teamRoster re-export the vocabulary they consume so importers
keep one natural import; roles.ts is the definition home. Also route
RosterMemberModal's owner-grant check through orgRoleForRole instead of a
re-derived `=== "instructor"` so "who becomes an org owner" lives in exactly
one place. No behavior change; tsc/eslint/prettier/vitest (1495) green.
